rename relationship to "ownership-by-file-overlap"
